Fixed basic code stuff
