+++ Makefile
@@ -1,0 +1,3 @@
+m:
+	g++ -Wall Moto.cpp Veiculo.cpp main.cpp -o main
+
