+++ Moto.cpp
@@ -1,0 +1,10 @@
+#include "Moto.hpp"
+
+Moto :: Moto(){
+    vel = 0;
+    blind = 0;
+    rodas = 2;
+    setTipo(1);
+    setvelMax(120);
+    setArma(false);
+}
