+++ Moto.hpp
@@ -1,0 +1,5 @@
+#include "Veiculo.cpp"
+class Moto : public Veiculo {
+    public: 
+        Moto();
+};
