+++ main.cpp
@@ -1,15 +1,12 @@
 #include <iostream>
-#include "classes.hpp"
+#include "Moto.cpp"
 
 using namespace std;
 
 int main (){
-
-
-    Moto *v1 = new Moto();
-
-    v1 -> imp();
-
-
+    // Moto *v1 = new Moto(); // gerando objeto por referencia
+    Moto m; // gerando objeto sem ser por referencia
+    m.imp();
+    // v1->imp();
     return 0;
 }
